requirements.txt
llama-index-core
llama-index-readers-file
llama-index-embeddings-huggingface
llama-index-vector-stores-chroma
llama-index-llms-gemini
llama-index-llms-ollama
sentence-transformers
chromadb
rank-bm25
fastapi
uvicorn
python-multipart
streamlit
python-dotenv
ragas
datasets
pytest
pypdf
requests
